requirements.txt
python-dotenv
requests
twilio
newsapi-python